feat: slider tooltip; tweaks n fixes for slider
